finish api implementation

+++ src/app/controllers/ContactController.js
@@ -1,5 +1,7 @@
 const { request, response } = require('express');
 const ContactsRepository = require('../repositories/ContactsRepository');
+const CategoriesRepository = require('../repositories/CategoriesRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -10,58 +12,95 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if(!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if(!contact) {
       return response.status(404).json({ error: 'Contact not found' });
     }
-
-    response.json(contact);
+    let category = null;
+    if (contact.category_id) {
+      category = await CategoriesRepository.findById(contact.category_id);
+    }
+    response.json({
+      contact,
+    });
   }
 
   async store(request, response) {
-    const {name, email, phone, category_id} = request.body;
+    const { name, email, phone, category_id } = request.body;
 
     if (!name) {
-      return response.status(400).json({error: 'Name is required'});
+      return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if(email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
 
-    if (contactExists){
-      return response.status(400).json({error: 'This e-mail is already in use'});
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
-    const contact = await ContactsRepository.create(name, email, phone, category_id);
+    if(category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category'});
+    }
 
-    response.json(contact);
+    const contact = await ContactsRepository.create(name, email || null, phone, category_id || null);
+
+    response.status(201).json({
+      contact,
+    });
   }
 
   async update(request, response) {
     const { id } = request.params;
     const {name, email, phone, category_id} = request.body;
 
+    if(!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
+
     const contactExists = await ContactsRepository.findById(id);
     if (!contactExists){
       return response.status(404).json({error: 'User not found'});
+    }
+
+    if(category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category'});
     }
 
     if (!name) {
       return response.status(400).json({error: 'Name is required'});
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
+    if(email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
 
-    if (contactByEmail && contactByEmail.id !== id ){
-      return response.status(400).json({error: 'This e-mail is already in use'});
+      if (contactByEmail && contactByEmail.id !== id ){
+        return response.status(400).json({error: 'This e-mail is already in use'});
+      }
     }
 
-    const contact = await ContactsRepository.update(id, {name, email, phone, category_id});
+    const contact = await ContactsRepository.update(id, {
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
+    });
     response.json(contact);
   }
 
   async delete(request,response) {
     const { id } = request.params;
+
+    if(!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id'});
+    }
 
     await ContactsRepository.delete(id);
     return response.status(204).end(  );
